Order table updated

+++ database/migrations/2020_10_30_192931_create_districts_table.php
@@ -16,13 +16,14 @@ class CreateDistrictsTable extends Migration
         Schema::create('districts', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->unsignedInteger('division_id');
+           $table->unsignedBigInteger('division_id');
             $table->timestamps();
 
 
               $table->foreign('division_id')
             ->references('id')->on('divisions')
             ->onDelete('cascade');
+
         });
     }
